requirements.txt
# Core
pydantic>=2.7
PyYAML>=6.0.1

# Workflow (multi-agent orchestration)
langgraph>=0.2.0

# LLM (OpenAI-compatible client; works with Gemini OpenAI-compat base_url)
openai>=1.40.0

# PDF text extraction (Channel A)
pymupdf>=1.24.0

# DOCX output
python-docx>=1.1.2

